requirements.txt
# Banco de dados
pymongo

# Configuração
python-dotenv

# API e HTTP
requests

# Web Framework
flask

# Análise de Dados
pandas==2.2.1
